Make the server listen on http, but redirect to https
I've added two protocol-specific sites and removed the default one.

The variables for the servers have been added to group_vars/forum.yml

There is a possibility of conflict with existing default configuraitons,
so there should be a check to see that these are the _only_ server configs
This is not implemented yet.

Fixed #2