Initial commit on categories branch
this branch will add the scripts and tests necessary to ensure that the
right categories are present.

This will be done using the [discourse API](https://docs.discourse.org),
and you'll need an API key to test it.

The categories will be tested with a few inspec tests and created
with a Ruby script.